Refactor mouse
