+++ src/mouse/mouse.pyi
